Add !voterocket
Adds the command "sm_voterocket", which allows players to vote on whether the main rocket class should be changed.
After enough players vote to change the main rocket class, a menu will display to all real, connected, and non-spectator players.
The menu will house all of the rocket classes defined in the dodgeball configurations files; selecting one will count towards a vote for which rocket class should be the main rocket class.
After every player with a menu chooses a rocket class (or no change), or, after 20 seconds elapses, the rocket class with the most votes will be set as the main rocket class.
The main rocket class is the only rocket class able to spawn. If there is no main rocket class defined in the dodgeball configurations files, any operation regarding a main rocket class wil be aborted.